philo

+++ philo/src/actions.c
@@ -6,7 +6,7 @@
 /*   By: dcologgi <dcologgi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/18 13:54:13 by dcologgi          #+#    #+#             */
-/*   Updated: 2023/05/18 15:26:41 by dcologgi         ###   ########.fr       */
+/*   Updated: 2023/05/23 14:13:58 by dcologgi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,17 +14,17 @@
 
 void	status(char *str, t_philo *philo)
 {
-	int	time;
+	uint64_t	time;
 
 	pthread_mutex_lock(&philo->table->write);
 	time = get_time(philo->table) - philo->table->start_time;
 	if (ft_strcmp(DIED, str) == 0 && philo->table->death == 0)
 	{
-		printf("%d %d %s\n", time, philo->id, str);
+		printf("%lu %d %s\n", time, philo->id, str);
 		philo->table->death = 1;
 	}
 	if (!philo->table->death)
-		printf("%d %d %s\n", time, philo->id, str);
+		printf("%lu %d %s\n", time, philo->id, str);
 	pthread_mutex_unlock(&philo->table->write);
 }
